we should not really create the logfile from the package
When starting snclient the logfile will be created.

Context:
We use systemd overrides to run snclient as regular user, for our
usecase this works just fine, but on rpm based distro's we still get the
`/var/log/snclient/snclient.log` file with root permissions because it
is created in the post install of the package.

Signed-off-by: BlackEagle <ike.devolder@gmail.com>