feature: cpu idle time in check_cpu_utilization
Add some more features in the counter class, it stores more values like
oldest, timesSet, retention and interval. Implement getFirst() and
checkRetention() functions and use them in the check_cpu_utilization.

Cpu utilization check can now detect if the maximum counter retention
time is too small for the query, or if the current counter interval
between earliest and the latest value is smaller than the interval. It
will also report and complain in the log if the counter search interval
being relatively smaller than the query interval.