revise external script discovery
External script discovery was using a glob with "*.*" as filter. This
did not match files without extensions, and under UNIX systems
executables can have no extension.

Instead of globbing, use filepath.WalkDir with a custom function. This
function checks if a file is a regular file and executable before adding
it to the list. Checking if a file is executable is only supported on
unix platforms. If a file is a symlink, it will be followed up until it
 reaches a non-symlink file, and only added if its within the
 scriptPath.

Normal search only adds files that are found directly under scriptPath
and not inside a subfolder. If the scriptPath ends with **, files in the
subdirectories are added as well.

When adding new scripts, their names are searched if they match an
existing check function.