fix check_files depth argument and windows paths
the maxDepth argument was confusing, as it did not include anything when
it was set to 0, clarify its usage better in the documentation. make
max-depth argument '0' and '1' include files/directories directly under
the path. now a search with max-depth=0 is valid and does not return
empty output

optimize the folder skipping, where if max-depth is specified and the
directories depth is exactly the max depth, it will be possibly added
and fs.SkipDir is immediately returned. previously it would only return
if depth>max depth, which increased search time a lot when searched from
a drive

move some of the early guards/returns around so that they can be called
earlier, move the needVersion variable as well. add comments about
return calls and their intended effect