improve check-files windows tests
fix the problem in the powershell code that generates relative paths
caused by the strict mode fix the depth calculation when base path is
directly a drive. Since the trailing slashes are taken out, "C:" and
"C:Windows" would have the same depth, but the folder should have higher
depth. Reintroduce the slash if the basePath is the drive so that
normalizing "C:" gives "C:\" and specifically handle those cases in
depth calculation with a depthCorrection variable add tests that search
for windows files directly under the drive