check_drivesize ignore errors if empty-state is set
fix the misunderstood condition about how drive/directory discovery
errors are handled. previously, if the drive was not found the errors
were always ignored, the check would not have any items in its list and
the empty-state override would function since the list is empty

corrected it so that the errors are ignored only when empty-state is
set. if empty-state is set, it will ignore discovery errors, produce and
empty list and the override will function. but if its not, it returns
the error just as before.